requirements.txt
pynput
pillow
pyperclip
